🤖 Refactor: props 타입 정리
